requirements.txt
mysql-connector-python
cachetools==4.1.1